+++ ui/app/adapters/application.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import RESTAdapter from 'ember-data/adapters/rest';
+import codesForError from '../utils/codes-for-error';
 
 const { get, computed, inject } = Ember;
 
@@ -21,8 +22,12 @@ export default RESTAdapter.extend({
 
   findAll() {
     return this._super(...arguments).catch(error => {
-      if (error.code === '501' || (error.errors && error.errors.findBy('status', '501'))) {
-        // Feature is not implemented in this version of Nomad
+      const errorCodes = codesForError(error);
+
+      const isNotAuthorized = errorCodes.includes('403');
+      const isNotImplemented = errorCodes.includes('501');
+
+      if (isNotAuthorized || isNotImplemented) {
         return [];
       }
